Support for DMA nodes, improved CPU nodes
Add support for DMA node. These transfer data using the Zynq's HP ports
directly to and from DDR memory. This improves throughput and CPU usage
at the expense of latency compared to the CPU node.

Driver framework extended to support per-node handlers, so that
DMA and CPU can co-exist.

Extend CPU node to support multiple instances.

Add ioctls to get/set routes through the fifo or cfg node directly, so
that the user need not be aware of the location of each node.