Allocate FIFO transfer buffers on heap. Fix race in release.
The transfer buffers on stack caused the following warning:
warning: the frame size of 1132 bytes is larger than 1024 bytes
Solve this by allocating the 1k transfer buffers on the heap.

When closing the fifo device, use the mutex that guards the open()
as well, to make sure these actions are atomic, in particular access
to "is_open" must be synchronized.