Rename READ/WRITE to FROM_BACKPLANE and TO_BACKPLANE
Using terms like "input" or "read" doesn't make it clear which way the
data is actually flowing. Use FROM and TO designations instead, which
alse makes it clear that the input/output count of the CPU node was
being read from the wrong registers, which wasn't a problem because
up until now these were invariably symmetric.