Allow varying the requested bytes in the from-logic buffers
If the buffer is resized, this makes it impossible for the user application
to request an odd multiple of 1k data for example. For to-logic data the
extra size is not a problem because it has to fill in the bytes_used anyway.

Change the from-logic interface to have "bytes_used" set to the requested
size before the request is made. This allows userpace to request sizes other
than those allocated.

Return EBUSY when a submitted block is already in transfer.