Add streaming DMA support
dma_alloc_coherent returns horrificly slow uncached memory on the Zynq. Take
matters into our own hands and use the streaming interface instead. This
appears to be exactly what the DMA API suggests.

The cause appeared to be that dma_common_mmap and dma_mmap_coherent set
pgprot_noncached on the returned mapping, thereby disabling the CPU cache
and crippling performance.

However, streaming DMA also doesn't perform well, because the calls to dma_sync
take up about the same amount of time as just copying the data. So the end
result is that the user must decide which mode to use, hence introducing a new
reconfiguration ioctl.

Apparently the only way to get decent performance on the Zynq is to use the
ACP instead of HP, so the hardware handles the coherency. When using the ACP
port, add "dma-coherent" to the devicetree to inform the system that the
hardware can enforce coherency. The drawback of the ACP interface is that
access to data outside the L2 cache (i.e. DDR RAM) is much slower than through
the HP port. But it's still faster than flushing caches in software.