Make DYPLO_IOC_TRESHOLD_QUERY/TELL symmetric
DYPLO_IOC_TRESHOLD_TELL sets the treshold (block size) for DMA nodes,
so it is much more logical that DYPLO_IOC_TRESHOLD_QUERY returns that
same value instead of the total buffer size.

To retrieve the size of the whole DMA buffer, an application can use
the DYPLO_IOC_DMA_RECONFIGURE call. With mode set to
DYPLO_DMA_MODE_RINGBUFFER_BOUNCE the call will not affect the system
when already in that mode, and it will return the size and number of
buffers.

This affects existing code, in particular the driver tests in libdyplo.