Implement DYPLO_IOCGSTATIC_ID
DYPLO_IOCGSTATIC_ID obtains the static ID of the current bitstream. This
can be used to assure that a partial to be programmed is compatible with
the current FPGA static contents.
The static ID is available in IP version 2015.1.4, if the IP is older,
the call will return EIO and set the ID to 0.

Display the static id in /proc/dyplo (may show as "0" when not
implemented).