Prevent cache conflict in DMA transfer to logic
When transferring data to logic, and the size of the transfers is not
cache aligned, both CPU and logic may be accessing the same cache line,
resulting in data corruption. Prevent this from happening by aligning
the "head" of the outgoing data pointer to something that it cache
aligned. In this case, simply align to page size, since that is also
guaranteed to be a power of two and a multiple of the cacheline size.