Release DMA blocks on close
When closing the DMA handle, release all resources. Usually a client would
do that through an ioctl call, but if the process was terminated abnormally,
the handle would only be closed. Without this change, an attempt to read
or write the dma in its default mode would fail because it would still be
in block transfer (a.k.a. zero-copy) mode after the unexpected close. So
there was no resource leak, as the next reset would release the blocks
anyway, but it did lead to unexpected EBUSY return values from read or
write calls.