Fix that changing to-logic DMA configuration interferes with from-logic
The dma_reset routine was for standalone mode only, which was using both
channels at the same time. The dyplo_dma_to_logic_block_free call already
resets the node, so the call to dma_reset is not desired or needed.

The effect was that when calling reconfigure on a to-logic DMA handle would
disrupt the from-logic stream as well.

Fixes: 5d4e134c ("Remove DMA standalone-mode code...")